+++ subterra-optim/src/main/java/io/toterra/subterra/optim/worldgen/ticking/package-info.java
@@ -1,0 +1,6 @@
+package io.toterra.subterra.optim.worldgen.ticking;
+
+/**
+ * worldgen tick layer: the ticking-chunk cache core (see
+ * {@link io.toterra.subterra.optim.worldgen.ticking.TickingChunkCache}).
+ */